[datasets/numeric] docs
