+++ docs/source/index_numeric.rst
@@ -1,0 +1,10 @@
+.. _ref-numeric:
+
+Numeric package (sktime.numeric)
+================================
+The *numeric* package contains numerical utilities.
+
+.. automodule:: sktime.numeric
+
+.. toctree::
+   :maxdepth: 1
