+++ sktime/numeric/__init__.py
@@ -1,27 +1,32 @@
-import numpy as np
+r"""
+.. currentmodule: sktime.numeric
 
+===============================================================================
+General numerical tools
+===============================================================================
 
-def mdot(*args):
-    """Computes a matrix product of multiple ndarrays
+.. autosummary::
+    :toctree: generated/
 
-    This is a convenience function to avoid constructs such as np.dot(A, np.dot(B, np.dot(C, D))) and instead
-    use mdot(A, B, C, D).
+    mdot
 
-    Parameters
-    ----------
-    *args : an arbitrarily long list of ndarrays that must be compatible for multiplication,
-        i.e. args[i].shape[1] = args[i+1].shape[0].
-    """
-    if len(args) < 1:
-        raise ValueError('need at least one argument')
-    args = list(args)[::-1]
-    x = args.pop()
-    i=0
-    while len(args):
-        y = args.pop()
-        try:
-            x = np.dot(x, y)
-            i+=1
-        except ValueError as ve:
-            raise ValueError(f'argument {i} and {i+1} are not shape compatible:\n{ve}')
-    return x
+===============================================================================
+Numerical tools for eigenvalue problems
+===============================================================================
+
+.. autosummary::
+    :toctree: generated/
+
+    eig_corr
+    sort_by_norm
+    spd_eig
+    spd_inv
+    spd_inv_split
+    spd_inv_sqrt
+    ZeroRankError
+
+"""
+
+from .utils import mdot
+from .eigen import eig_corr, sort_by_norm, spd_eig, spd_inv, spd_inv_split, spd_inv_sqrt, ZeroRankError
+
